requirements.txt
huaweicloudsdkcore==3.1.*
huaweicloudsdkwaf==3.1.*
python-dotenv==1.0.1